refactor: AbsFileRepository, StaticFileRepository refactor: AbsFileRepository to interface refactor: StaticFileRepository.append return null when not exist file add: add test append on not exist file
